Test name refactor, little code cleanup

+++ src/main/java/PaddleDrawer.java
@@ -9,6 +9,6 @@ public class PaddleDrawer implements Drawer {
     }
 
     public void draw(Graphics2D g2){
-        g2.fillRect((int)paddle.getxPosition(), (int)paddle.getyPosition(), paddle.getPADDLE_WIDTH(), paddle.getPADDLE_HEIGHT());
+        g2.fillRect(paddle.getxPosition(), paddle.getyPosition(), paddle.getPADDLE_WIDTH(), paddle.getPADDLE_HEIGHT());
     }
 }
